Start web UI (#15)
* Start web UI

Gets enough up to show the most recent points. Can iterate from here

* Fix up distribution

* use time rather than points

* switch to flexbox, inline the data as geojson

flexbox is the new cool for layouts

move to geojson for data, as it's a standard type. inline it in the document,
so we can do a bunch of the form logic in go server side. this is easier for
me

* inline css/js, use jquery ui date picker

not enough css and js to justify external files

use jquery ui for datepicker in safari.

* Make date filter work

* Add accuracy filtering

Useful for removing really vague points when viewing. Will be
even more useful for plotting lines when we get to that

* Add a popup with a little info