Disable spatialite, WAL by default
spatialite has proven to be a continual pain in the ass for something we
don't actually use in the app, so drop it for now. Can continue to use it
in the CLI for ad-hoc if we want.

Enable WAL mode by default. This ruins the 'one file' thing, but also is
what should be used in a serving environment