Upgrade dependencies, Use GitHub Registry, Database usage updates (#32)
* Upgrade to 1.18, use newer docker build stuff

The new stuff automatics tags and things like that

* Disable spatialite, WAL by default

spatialite has proven to be a continual pain in the ass for something we
don't actually use in the app, so drop it for now. Can continue to use it
in the CLI for ad-hoc if we want.

Enable WAL mode by default. This ruins the 'one file' thing, but also is
what should be used in a serving environment

* Use os/gh actions lint, normal sqlite build

* Drop cache mode, like we did in #30

https://github.com/mattn/go-sqlite3/issues/632#issue-360437494

* Set a busy timeout

* Remove DB write mutex

we shouldn't actually need that, esp. with wal mode + busy timeout etc.

* Upgrade go version / sqlite driver again

* Tests use real DB on disk too

rather than potentially masking over db config issues by using in-mem

* Drop transaction around venue sync

we ran that in one big tx, which took a couple seconds and locked the DB. I
can't think of any real reason why it can't be an incremental update, the writes
are all upserts. So drop the TX

* upgrade actions to 1.19 too

* Fix lint errors