Set env secret on tpsync too
We removed the env default from the flag, but then only set the env var on the
web server. This sets it on the tpsync instance too if needed.

The multiple copies of the secret is a bit jank so might refactor it a bit, now
is not that time tho.